No type and lint errors remaining.
