+++ src/Neuralm.Presentation.Web/src/.env
@@ -1,0 +1,1 @@
+API_URL = "https://Neuralm.net"
