--- src/Neuralm.Presentation.Web/src/store/Default.store.ts
@@ -1,6 +1,0 @@
-/**
- * Represents the default interface for a Vuex store.
- */
-export default interface IStore {
-  namespaced: boolean;
-}
