+++ src/Neuralm.Presentation.Web/tslint.json
@@ -14,6 +14,27 @@
     "no-consecutive-blank-lines": false,
     "object-literal-sort-keys": false,
     "ordered-imports": false,
-    "quotemark": [true, "single"]
+    "quotemark": [true, "single"],
+    "trailing-comma": false,
+    "member-ordering": [
+      true, {
+      "order": [
+        "private-static-field",
+        "private-instance-field",
+        "public-static-field",
+        "public-instance-field",
+        "public-constructor",
+        "private-constructor",
+        "public-instance-method",
+        "protected-instance-method",
+        "private-instance-method"
+        ]
+      }
+    ],
+    "variable-name": {
+      "options": [
+        "allow-leading-underscore"
+      ]
+    }
   }
 }
